move default roles seeding into create_workspace handler
