--- .sqlx/query-23772dd7728bcacefc24dfa0e44c829f2fc6b2ee2127339f4da2293620fad3d1.json
@@ -1,20 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "INSERT INTO resource(id, resource_type_id, owner_id, created) VALUES (GENERATE_RESOURCE_ID(), (SELECT id FROM resource_type WHERE name = 'workspace'), gen_random_uuid(), NOW()) RETURNING id;",
-  "describe": {
-    "columns": [
-      {
-        "ordinal": 0,
-        "name": "id",
-        "type_info": "Uuid"
-      }
-    ],
-    "parameters": {
-      "Left": []
-    },
-    "nullable": [
-      false
-    ]
-  },
-  "hash": "23772dd7728bcacefc24dfa0e44c829f2fc6b2ee2127339f4da2293620fad3d1"
-}
