+++ .sqlx/query-30780fb92a5d8c3d78e7284a0612428d3f73027b50b8573fed22ebd011727925.json
@@ -1,0 +1,20 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "INSERT INTO resource(id, resource_type_id, owner_id, created) VALUES (GENERATE_RESOURCE_ID(), (SELECT id FROM resource_type WHERE name = 'workspace'), gen_random_uuid(), NOW()) RETURNING id AS \"id: Uuid\";",
+  "describe": {
+    "columns": [
+      {
+        "ordinal": 0,
+        "name": "id: Uuid",
+        "type_info": "Uuid"
+      }
+    ],
+    "parameters": {
+      "Left": []
+    },
+    "nullable": [
+      false
+    ]
+  },
+  "hash": "30780fb92a5d8c3d78e7284a0612428d3f73027b50b8573fed22ebd011727925"
+}
