+++ .sqlx/query-4e2a186ec7b34027195a6800df7a36dbbde1043bd2c894fd539c8947bfedaeee.json
@@ -51,7 +51,9 @@
       true,
       false,
       false,
-      true
+      true,
+      false,
+      false
     ]
   },
   "hash": "4e2a186ec7b34027195a6800df7a36dbbde1043bd2c894fd539c8947bfedaeee"
