+++ .sqlx/query-ee6a287cd21311680135dca639cc93f49ed8b3d9081ce8e4e3f678b3491ac1c0.json
@@ -52,7 +52,9 @@
       true,
       false,
       false,
-      true
+      true,
+      false,
+      false
     ]
   },
   "hash": "ee6a287cd21311680135dca639cc93f49ed8b3d9081ce8e4e3f678b3491ac1c0"
